Update README.md
The quiz app is a project for the rounding up of my android development course (android basic track). 
The application simply contains geographical question making use of a linear layout and also different views ranging from TextView, Button, RadioButton and so on.
The users is expected to attempt all the question provided in the app and then submit to see their final scores.